Test scaffolding and conditionally require uploads
