+++ src/applications/disability-benefits/all-claims/constants.js
@@ -137,6 +137,8 @@ export const DATA_PATHS = {
     'view:hasEvidenceFollowUp.view:selectableEvidenceTypes.view:hasVAMedicalRecords',
   hasPrivateEvidence:
     'view:hasEvidenceFollowUp.view:selectableEvidenceTypes.view:hasPrivateMedicalRecords',
+  hasPrivateRecordsToUpload:
+    'view:uploadPrivateRecordsQualifier.view:hasPrivateRecordsToUpload'
 };
 
 export const VA_FORM4142_URL = 'https://www.vba.va.gov/pubs/forms/VBA-21-4142-ARE.pdf';
